requirements.txt
discord.py[voice]>=2.4.0
yt-dlp
google-api-python-client==2.110.0
PyNaCl==1.5.0
python-dotenv
audioop-lts; python_version >= "3.13"
aiohttp
edge-tts
flask
pytest
pytest-asyncio
aioresponses